added implementation for Application and Api layers, add some tests

+++ Application/UseCases/Commands/Rent/CompleteRent/CompleteRentHandler.cs
@@ -11,19 +11,21 @@ public class CompleteRentHandler(
     IUnitOfWork unitOfWork,
     TimeProvider timeProvider) : IRequestHandler<CompleteRentCommand, Result<CompleteRentResponse>>
 {
-    public async Task<Result<CompleteRentResponse>> Handle(CompleteRentCommand command, CancellationToken cancellationToken)
+    public async Task<Result<CompleteRentResponse>> Handle(
+        CompleteRentCommand command,
+        CancellationToken cancellationToken)
     {
         var rent = await rentRepository.GetById(command.RentId);
         if (rent == null) return Result.Fail(new NotFound("Rent not found"));
-        
+
         var actualAmount = rent.Complete(timeProvider);
-        
+
         rentRepository.Update(rent);
 
         var commitResult = await unitOfWork.Commit();
 
         return commitResult.IsSuccess
-            ? new CompleteRentResponse((double)actualAmount)
+            ? new CompleteRentResponse(actualAmount)
             : commitResult;
     }
 }
