+++ Application/UseCases/Commands/Rent/CompleteRent/CompleteRentResponse.cs
@@ -1,3 +1,3 @@
 namespace Application.UseCases.Commands.Rent.CompleteRent;
 
-public record CompleteRentResponse(double ActualAmount);
+public record CompleteRentResponse(decimal ActualAmount);
